refactor(#190): gallery 및 저장공간 권한 관련 부분 클래스로 분리
