+++ AOS/app/src/main/java/boostcamp/and07/mindsync/ui/util/ImagePickerHandler.kt
@@ -1,0 +1,113 @@
+package boostcamp.and07.mindsync.ui.util
+
+import android.Manifest
+import android.content.Intent
+import android.content.pm.PackageManager
+import android.net.Uri
+import android.os.Build
+import android.provider.MediaStore
+import androidx.activity.result.PickVisualMediaRequest
+import androidx.activity.result.contract.ActivityResultContracts
+import androidx.appcompat.app.AppCompatActivity
+import androidx.core.app.ActivityCompat
+import androidx.core.content.ContextCompat
+import androidx.fragment.app.FragmentActivity
+
+class ImagePickerHandler(
+    private val activity: FragmentActivity,
+    private val onImageSelected: (Uri) -> Unit,
+) {
+    private val imagePickerLauncher =
+        activity.registerForActivityResult(ActivityResultContracts.PickVisualMedia()) { uri ->
+            uri?.let { onImageSelected(it) }
+        }
+
+    fun launchImagePicker() {
+        imagePickerLauncher.launch(PickVisualMediaRequest(ActivityResultContracts.PickVisualMedia.ImageOnly))
+    }
+
+    fun checkPermissionsAndLaunchImagePicker() {
+        when {
+            ContextCompat.checkSelfPermission(
+                activity,
+                getPermissionReadMediaImagesOrReadExternalStorage(),
+            ) == PackageManager.PERMISSION_GRANTED -> {
+                launchImagePicker()
+            }
+
+            activity.shouldShowRequestPermissionRationale(
+                getPermissionReadMediaImagesOrReadExternalStorage(),
+            ) -> {
+                requestGalleryPermission()
+            }
+
+            else -> {
+                requestGalleryPermission()
+            }
+        }
+    }
+
+    private fun getPermissionReadMediaImagesOrReadExternalStorage() =
+        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.TIRAMISU) {
+            Manifest.permission.READ_MEDIA_IMAGES
+        } else {
+            Manifest.permission.READ_EXTERNAL_STORAGE
+        }
+
+    private fun requestGalleryPermission() {
+        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.TIRAMISU) {
+            galleryPermissionLauncher.launch(Manifest.permission.READ_MEDIA_IMAGES)
+        } else {
+            checkPermissionAndLaunchImageSelector()
+        }
+    }
+
+    private val galleryPermissionLauncher =
+        activity.registerForActivityResult(ActivityResultContracts.RequestPermission()) { isGranted ->
+            if (isGranted) {
+                launchImagePicker()
+            }
+        }
+
+    private val imageResult =
+        activity.registerForActivityResult(
+            ActivityResultContracts.StartActivityForResult(),
+        ) { result ->
+            if (result.resultCode == AppCompatActivity.RESULT_OK) {
+                result.data?.data?.let { uri ->
+                    onImageSelected(uri)
+                }
+            }
+        }
+
+    private fun checkPermissionAndLaunchImageSelector() {
+        val readPermission =
+            ContextCompat.checkSelfPermission(activity, Manifest.permission.READ_EXTERNAL_STORAGE)
+
+        if (readPermission == PackageManager.PERMISSION_GRANTED) {
+            launchImageSelector()
+        } else {
+            requestExternalStoragePermission()
+        }
+    }
+
+    private fun launchImageSelector() {
+        val intent =
+            Intent(Intent.ACTION_PICK).apply {
+                setDataAndType(MediaStore.Images.Media.EXTERNAL_CONTENT_URI, "image/*")
+            }
+        imageResult.launch(intent)
+    }
+
+    private fun requestExternalStoragePermission() {
+        ActivityCompat.requestPermissions(
+            activity,
+            arrayOf(Manifest.permission.READ_EXTERNAL_STORAGE),
+            REQUEST_CODE_PERMISSIONS,
+        )
+    }
+
+    companion object {
+        const val REQUEST_CODE_PERMISSIONS = 1
+    }
+}
